fix(main): put rewards in account store
